st/nine: Implement ps3 advanced input definition feature
ps3 allows definitions of the inputs like:
DCL_TEXCOORD0 v0.xy;
DCL_NORMAL2 v0.z;
DCL_NORMAL3 v0.w;

Nine wouldn't have handled this situation properly.

Apparently very few applications use this feature.

Still remain an issue with this new implementation:
It is allowed to do indirect addressing on the ps inputs.

Since here the inputs are not contiguous (we allocate temps)
it cannot be implemented (we have an assert for that currently
in the code, and at least one app was reported to need this to
work)

Signed-off-by: Axel Davy <axel.davy@ens.fr>