st/nine: Rework DSA state handling
Previous mechanism to update state was:
. Store render state values, and track down
which render states are changed
. At draw time fill the pipe structures that need updating
and commit the changes

This patch introduces another way:
. Store render state values
. Cache the pipe structures.
. When a render state value is changed,
if it is a simple change, update directly the
pipe structure. Else put a flag so that the structure
is entirely update at draw time
. At draw time the pipe structures are commited.

This is retrocompatible with the previous way, which is still used
for stateblocks and render states that would need to modify to many parts of
the structure (when there is a lot of interdependance between the states).

This patch only implements it for DSA states.

Signed-off-by: Axel Davy <axel.davy@ens.fr>